requirements.txt
boto3
python-dotenv
pylint
pytest
pytest-cov